+++ app/src/main/java/org/breezyweather/main/adapters/main/holder/HeaderViewHolder.kt
@@ -126,7 +126,7 @@ class HeaderViewHolder(parent: ViewGroup, weatherView: WeatherView) : AbstractMa
             detailDisplayList.forEach { detailDisplay ->
                 detailDisplay.getCurrentValue(LocalContext.current, current, isDaylight)?.let {
                     if (!firstItem) {
-                        Divider(color = Color.White.copy(alpha = 0.5f), thickness = 0.5.dp)
+                        HorizontalDivider(color = Color.White.copy(alpha = 0.5f), thickness = 0.5.dp)
                     } else {
                         firstItem = false
                     }