+++ app/src/main/java/org/breezyweather/settings/activities/WorkerInfoActivity.kt
@@ -36,7 +36,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.fastForEach
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.asFlow
 import androidx.work.WorkInfo
 import androidx.work.WorkQuery
 import dagger.hilt.android.lifecycle.HiltViewModel