+++ app/src/main/java/org/breezyweather/sources/smhi/SmhiService.kt
@@ -1,0 +1,65 @@
+/**
+ * This file is part of Breezy Weather.
+ *
+ * Breezy Weather is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, version 3 of the License.
+ *
+ * Breezy Weather is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.breezyweather.sources.smhi
+
+import android.content.Context
+import android.graphics.Color
+import io.reactivex.rxjava3.core.Observable
+import org.breezyweather.common.basic.models.Location
+import org.breezyweather.common.basic.wrappers.WeatherWrapper
+import org.breezyweather.common.source.HttpSource
+import org.breezyweather.common.source.MainWeatherSource
+import org.breezyweather.common.source.SecondaryWeatherSourceFeature
+import retrofit2.Retrofit
+import javax.inject.Inject
+
+class SmhiService @Inject constructor(
+    client: Retrofit.Builder
+) : HttpSource(), MainWeatherSource {
+
+    override val id = "smhi"
+    override val name = "SMHI"
+    override val privacyPolicyUrl = "https://www.smhi.se/omsmhi/hantering-av-personuppgifter/hantering-av-personuppgifter-1.135429"
+
+    override val color = Color.rgb(0, 0, 0)
+    override val weatherAttribution = "SMHI (Creative commons Erkännande 4.0 SE)"
+
+    private val mApi by lazy {
+        client
+            .baseUrl(SMHI_BASE_URL)
+            .build()
+            .create(SmhiApi::class.java)
+    }
+
+    override val supportedFeaturesInMain = listOf<SecondaryWeatherSourceFeature>()
+
+    override fun requestWeather(
+        context: Context, location: Location,
+        ignoreFeatures: List<SecondaryWeatherSourceFeature>
+    ): Observable<WeatherWrapper> {
+        val smhiForecastResult = mApi.getForecast(
+            location.longitude,
+            location.latitude
+        )
+
+        return smhiForecastResult.map { convert(it, location.timeZone) }
+    }
+
+    companion object {
+        private const val SMHI_BASE_URL = "https://opendata-download-metfcst.smhi.se/api/"
+    }
+}