+++ app/src/main/java/org/breezyweather/sources/smhi/json/SmhiTimeSeries.kt
@@ -14,26 +14,14 @@
  * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.breezyweather.db.entities
+package org.breezyweather.sources.smhi.json
 
-import io.objectbox.annotation.Entity
-import io.objectbox.annotation.Id
-import java.util.*
+import kotlinx.serialization.Serializable
+import org.breezyweather.common.serializer.DateSerializer
+import java.util.Date
 
-/**
- * Alert entity.
- *
- * [Alert]
- */
-@Entity
-class AlertEntity(
-    @field:Id var id: Long = 0,
-    var formattedId: String,
-    var alertId: String,
-    var startDate: Date? = null,
-    var endDate: Date? = null,
-    var description: String,
-    var content: String? = null,
-    var priority: Int,
-    var color: Int
+@Serializable
+data class SmhiTimeSeries(
+    @Serializable(DateSerializer::class) val validTime: Date,
+    val parameters: List<SmhiParameter>
 )